refactor(workflows): neutralize automation run service filename
