+++ turbo/apps/api/src/signals/services/zero-workflow-automation-poller.service.ts
@@ -14,7 +14,7 @@ import { tapError } from "../utils";
 import { dispatchFailedRunCallbacks } from "./agent-run-callback.service";
 import { workflowAutomationColumns } from "./autonomy-budget-schema.service";
 import { calculateNextRun } from "./time-automation";
-import { runWorkflowAutomationNow$ } from "./zero-workflow-automation-run.service";
+import { runWorkflowAutomationNow$ } from "./workflow-automation-run.service";
 import {
   scheduleTriggerContext,
   type DueWorkflowAutomation,
