+++ turbo/apps/api/src/signals/services/zero-workflow-automation.service.ts
@@ -122,7 +122,7 @@ import {
   revealWorkflowWebhookSecretFields,
 } from "./workflow-webhook-automation.service";
 import { dispatchFailedRunCallbacks } from "./agent-run-callback.service";
-import { runWorkflowAutomationNow$ } from "./zero-workflow-automation-run.service";
+import { runWorkflowAutomationNow$ } from "./workflow-automation-run.service";
 import type { RunWorkflowAutomationResult } from "./zero-workflow-automation-launch.service";
 import { manualTriggerSource } from "./workflow-automation-trigger-source";
 import {
